development
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@105 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/army.h
@@ -65,7 +65,7 @@ namespace Army
 	Dialog::answer_t ShowDialogInfo(const Heroes * heroes = NULL, bool quickshow = false);
 
     private:
-        std::pair<Monster::monster_t, u8> pair;
+        std::pair<Monster::monster_t, u16> pair;
     };
 
     class SelectBar
